+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+Changes:
+
+- Fix selected colour in colour changing dialog
+
 ## v2.0.4 (2021-07-27)
 
 Changes: